app/requirements/base.txt
Django==5.0.1
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
mysqlclient==2.2.1